+++ R/sanitize_vcov.R
@@ -10,21 +10,35 @@ sanitize_vcov <- function(vcov, number_of_models, ...) {
     vcov <- ellip[["statistic_override"]]
   }
 
+  if (is.null(vcov)) {
+    vcov <- list(vcov)
+  }
+
+  # corner case: `vcov = list(NULL)`
+  if (isTRUE(checkmate::check_list(vcov)) &&
+      length(vcov) == 1 &&
+      is.null(vcov[[1]])) {
+    vcov <- rep(list(NULL), number_of_models)
+  }
+
   # recycling vcov for multiple models
   # single string/formulas/matrices/functions: apply to every model
-  if (is.null(vcov) ||
-      isTRUE(checkmate::check_formula(vcov)) ||
-      isTRUE(checkmate::check_matrix(vcov))  ||
-      isTRUE(checkmate::check_function(vcov)) ||
-      isTRUE(checkmate::check_numeric(vcov)) ||
-      isTRUE(checkmate::check_character(vcov, len = 1))) {
-    vcov <- rep(list(vcov), number_of_models)
-  # character vector to list (can be any length for single model many vcov)
-  } else if (isTRUE(checkmate::check_character(vcov))) {
-    vcov <- as.list(vcov)
-  # check first class because some models inherit from "list"
-  } else if (class(vcov)[1] != "list") {
-    stop("The value of the `vcov` argument is invalid. Please refer to the documentation.")
+  if (!isTRUE(checkmate::check_list(vcov))) {
+    if (is.null(vcov) ||
+        isTRUE(checkmate::check_formula(vcov)) ||
+        isTRUE(checkmate::check_matrix(vcov))  ||
+        isTRUE(checkmate::check_function(vcov)) ||
+        # checkmate oddity for list(NULL) which flags as numeric
+        isTRUE(checkmate::check_numeric(vcov)) ||
+        isTRUE(checkmate::check_character(vcov, len = 1))) {
+      vcov <- rep(list(vcov), number_of_models)
+    # character vector to list (can be any length for single model many vcov)
+    } else if (isTRUE(checkmate::check_character(vcov))) {
+      vcov <- as.list(vcov)
+    # check first class because some models inherit from "list"
+    } else if (class(vcov)[1] != "list") {
+      stop("The value of the `vcov` argument is invalid. Please refer to the documentation.")
+    }
   }
 
   checkmate::check_true(length(vcov) == number_of_models)