+++ tests/testthat/test-modelsummary_get.R
@@ -4,7 +4,7 @@ mod <- lm(Girth ~ Height + Volume, data = trees)
 test_that("options(modelsummary_get)", {
     options(modelsummary_get = "broom")
     tab1 <- get_gof(mod)
-    expect_equal(ncol(tab1), 13)
+    expect_equal(ncol(tab1), 14)
 
     options(modelsummary_get = "easystats")
     tab2 <- get_gof(mod)